refactor : PostCategoryDto 리팩토링
