+++ src/main/java/com/smile/petpat/post/category/repository/querydsl/CategoryRepositoryQuerydsl.java
@@ -1,9 +1,9 @@
 package com.smile.petpat.post.category.repository.querydsl;
 
-import com.smile.petpat.post.category.dto.PostCategoryInfo;
+import com.smile.petpat.post.category.dto.PostCategoryDto;
 
 import java.util.List;
 
 public interface CategoryRepositoryQuerydsl {
-    List<PostCategoryInfo.TradeCategoryDetailResponse> getTradeCategoryAndCnt(Long tradeCategoryId);
+    List<PostCategoryDto.TradeCategoryDetailResponse> getTradeCategoryAndCnt(Long tradeCategoryId);
 }
